+++ selfdrive/test/process_replay/camera_replay.py
@@ -5,7 +5,7 @@ import time
 from typing import Any
 from tqdm import tqdm
 
-from common.android import ANDROID
+from common.hardware import ANDROID
 os.environ['CI'] = "1"
 if ANDROID:
   os.environ['QCOM_REPLAY'] = "1"
@@ -84,6 +84,8 @@ if __name__ == "__main__":
 
   if update:
     ref_commit = get_git_commit()
+    if ref_commit is None:
+      raise Exception("couldn't get ref commit")
     log_fn = "%s_%s_%s.bz2" % (TEST_ROUTE, "model", ref_commit)
     save_log(log_fn, log_msgs)
     with open("model_replay_ref_commit", "w") as f:
@@ -101,4 +103,3 @@ if __name__ == "__main__":
       f.write(diff2)
 
     sys.exit(int(failed))
-