+++ selfdrive/test/test_car_models.py
@@ -433,6 +433,14 @@ routes = {
     'carFingerprint': MAZDA.CX5,
     'enableCamera': True,
   },
+  "10b5a4b380434151|2020-08-26--17-11-45": {
+    'carFingerprint': MAZDA.CX9,
+    'enableCamera': True,
+  },
+  "74f1038827005090|2020-08-26--20-05-50": {
+    'carFingerprint': MAZDA.Mazda3,
+    'enableCamera': True,
+  },
 }
 
 passive_routes: List[str] = [
@@ -442,8 +450,12 @@ forced_dashcam_routes = [
   # Ford fusion
   "f1b4c567731f4a1b|2018-04-18--11-29-37",
   "f1b4c567731f4a1b|2018-04-30--10-15-35",
-  # Mazda
+  # Mazda CX5
   "32a319f057902bb3|2020-04-27--15-18-58",
+  # Mazda CX9
+  "10b5a4b380434151|2020-08-26--17-11-45",
+  # Mazda3
+  "74f1038827005090|2020-08-26--20-05-50",
 ]
 
 # TODO: add routes for these cars