+++ selfdrive/ui/qt/window.hpp
@@ -7,6 +7,7 @@
 #include <QOpenGLFunctions>
 #include <QStackedLayout>
 
+#include "qt/qt_sound.hpp"
 #include "ui/ui.hpp"
 
 class MainWindow : public QWidget
@@ -25,7 +26,12 @@ public slots:
 
 };
 
-
+#ifdef QCOM2
+const int vwp_w = 2160;
+#else
+const int vwp_w = 1920;
+#endif
+const int vwp_h = 1080;
 class GLWindow : public QOpenGLWidget, protected QOpenGLFunctions
 {
   Q_OBJECT
@@ -45,7 +51,9 @@ protected:
 private:
   QTimer * timer;
   UIState * ui_state;
-  pthread_t connect_thread_handle;
+  QtSound sound;
+
+  bool onroad = true;
 
 public slots:
   void timerUpdate();